pyproject.toml
[project]
name = "foundry"
version = "1.0.0"
description = "Generate Claude Code project folders from reusable building blocks"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
keywords = ["claude", "ai", "project-generation", "scaffolding"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Code Generators",
]
dependencies = [
    "PySide6",
    "PyYAML",
    "Jinja2",
    "pydantic",
]

[project.urls]
Homepage = "https://github.com/gregg/foundry"
Repository = "https://github.com/gregg/foundry"

[project.scripts]
foundry = "foundry_app.main:main"
foundry-cli = "foundry_app.cli:cli_main"

[dependency-groups]
dev = [
    "pytest",
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["foundry_app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
